+++ cratedb_toolkit/testing/testcontainers/util.py
@@ -12,25 +12,12 @@
 #    under the License.
 import logging
 import os
-from typing import Any
 
 from testcontainers.core.container import DockerContainer
 
+from cratedb_toolkit.util.data import asbool
+
 logger = logging.getLogger(__name__)
-
-
-# from sqlalchemy.util.langhelpers
-# from paste.deploy.converters
-def asbool(obj: Any) -> bool:
-    if isinstance(obj, str):
-        obj = obj.strip().lower()
-        if obj in ["true", "yes", "on", "y", "t", "1"]:
-            return True
-        elif obj in ["false", "no", "off", "n", "f", "0"]:
-            return False
-        else:
-            raise ValueError("String is not true/false: %r" % obj)
-    return bool(obj)
 
 
 class ExtendedDockerContainer(DockerContainer):